move gui-editor.service, sidebar.type and gui-editor.event.ts to shared
