--- client/src/app/components/gui-editor/gui-editor.events.ts
@@ -1,11 +1,0 @@
-import {SidebarType} from "./sidebar/editor-sidebar.component";
-
-export enum GuiEditorEventType {
-    ShowSidebar
-}
-
-export interface ShowSidebarEvent {
-    data: {
-        sidebarType: SidebarType
-    }
-}
