+++ client/src/app/components/gui-editor/shared/gui-editor.events.ts
@@ -1,0 +1,7 @@
+import {SidebarType} from "../sidebar/editor-sidebar.component";
+
+export interface ShowSidebarEvent {
+    data: {
+        sidebarType: SidebarType
+    }
+}
