+++ client/src/app/components/gui-editor/shared/gui-editor.service.ts
