+++ client/src/app/components/gui-editor/shared/sidebar.types.ts
@@ -1,0 +1,1 @@
+export type SidebarType = "expression" | "editor";
