+++ client/src/app/components/gui-editor/sidebar/editor-sidebar.component.ts
@@ -8,17 +8,13 @@ import {
     trigger,
     Input
 } from "@angular/core";
-import {GuiEditorService} from "../gui-editor.service";
+import {GuiEditorService} from "../shared/gui-editor.service";
 import {VisibilityState} from "../animation.states";
-import {ShowSidebarEvent} from "../gui-editor.events";
+import {ShowSidebarEvent} from "../shared/gui-editor.events";
 import {BehaviorSubject} from "rxjs/Rx";
+import {SidebarType} from "../shared/sidebar.types.ts"
 
 require ("./editor-sidebar.component.scss");
-
-export enum SidebarType {
-    expression,
-    editor
-}
 
 /** TODO: make this switch between an expression editor and an object inspector*/
 @Component({
@@ -26,13 +22,13 @@ export enum SidebarType {
     animations: [
         trigger("sidebarState", [
             state("visible", style({
-                width: '40%',
-                display: 'block',
+                width:"40%",
+                display: "block",
                 overflowY: "auto"
             })),
             state("hidden", style({
-                width: '10%',
-                display: 'none',
+                width: "10%",
+                display: "none",
                 overflowY: "hidden"
             })),
             transition("hidden => visible", animate("100ms ease-in")),
@@ -46,6 +42,13 @@ export enum SidebarType {
                         </div>
                         <div class="sidebar-content">
                             <!-- TODO: add expression and objects editor -->
+                            <span *ngIf="sidebarType === 'editor'">
+                                Editor
+                            </span>
+                            
+                            <span *ngIf="sidebarType === 'expression'">
+                                Expression
+                            </span>
                             This is the right sidebar content
                         </div>
                     </div>
@@ -59,9 +62,12 @@ export class EditorSidebarComponent implements OnInit {
     /** State of the sidebar animation */
     private sidebarState: VisibilityState;
 
+    private sidebarType: SidebarType;
+
     constructor(private guiEditorService: GuiEditorService) {
         this.guiEditorService.publishedSidebarEvents.subscribe((event: ShowSidebarEvent) => {
-            this.showSideBar(event.data.sidebarType);
+            this.sidebarType = event.data.sidebarType;
+            this.showSideBar();
         });
     }
 
@@ -71,7 +77,7 @@ export class EditorSidebarComponent implements OnInit {
         });
     }
 
-    showSideBar(sidebarType: SidebarType): void {
+    showSideBar(): void {
         this.sidebarVisibility.next("visible");
     }
